Added sanity check function.
The function is meant to be called in every other function taking E164
as an input argument before doing anything else to the argument value.
In this way we try ensure the value represents a valid number, in
order to lower possibility of working with a corrupted value and help
catch possible internal representation bugs early.